requirements.txt
gtts
pandas
scikit-learn
numpy
scikit-build
opencv-python
pillow
scikit-image
pytesseract
easyocr
imutils
matplotlib
moviepy
nltk
vosk
pydub
transformers
torch
sacremoses